Save Post File on Aborted Push to Blog
If the attempt to push the file to the blog fails for some reason, then the file
is saved rather than lost entirely.  This is nice when blogtool it being
utilized without a filename in the command line so that a post is no longer lost
entirely.

Some minor bug fixes as well.