Comment Option Requires PostID Arg
Since posting a comment always requires a postid, I've added it as a required
argument to the option.

This necessitated a fair amount of monkeying with header building and the like.
Overall, it's been improved but may need more work.  There are several
permutations depending on whether specific blogs are specified and how many
blogs are setup in the rc file and whether a comment or post is being published.