HTML2MD Fixes
Added the `method="html"` option to the `tostring` calls to make sure that tags
with no text are generated with a proper closing tag rather than just a `/>`.

Before change, `iframe` tags would be returned from `tostring` as:

    <iframe attribute="something" attribute2="somethingelse"/>

This is because the source HTML has no text between the start and end tags.

After change, `iframe` tags are returned as:

    <iframe attribute="something" attribute2="somethingelse"></iframe>

HTML code in a browser doesn't take kindly to the first one.

Besides, we *want* HTML output from the string conversion.