Added .gitattributes
There is path information we don't want to change in the master branch for the
xmlproxy/__init__.py file.  When merging with dev though, it will always change
because dev is path'd to use the local files, not the system.  The master branch
is path'd as if for a system install.