Charset Option now Valid for Comment Editting
After adding the charset option, comment editting was broken because the
fileprocessor now requires a charset to be established.  Now, the charset can be
set on the command line which will set the charset for comment editting.
Unfortunately, this created a condition where the editor would be launched for
writing a post after editting a comment.  I had to modify the runeditor checking
in blogtool.py to account for the new condition.  I commented the code as well
for future reference.