Reorganized for publishing to pypi
Reorganized project for eventual publication to pypi.  In addition to moving
files around a bit, I had to modify a number of them to maintain funcitionality.
With the exception of `xmlproxy/__init__.py`, the changes were minor.  It was an
overall improvement as some path dependent stuff is now gone.