Escape certain characters during GetPost
The lxml parser doesn't like certain characters unescaped, like the '&'.
Normally, the XMLRPC exchange causes these characters to be properly escaped,
for instance '&' as '&amp;'.  But there are certain instances where they are not
properly escaped.  We now search for the '&' characters in the text stream and
escape them prior to parsing.

Further, the `tostring` function provided by lxml will not convert the escaped
pattern back to their single character, i.e. change '&amp;' to '&'.  The code
properly handles this situation.

Note that we should probably eliminate all usage of the `tostring` function and
create our own serializer to deal with this situation.