Now support EXCERPT keyword in Header
Have added support for creating an excerpt for a blog post, which necessitated
the addition of the EXCERPT keyword for headers.  To more fully support what the
excerpt in the header, I've provided for python-style docstring quoting.  These
strings are delimited by a sequence of triple double-quotes, like the docstrings
in python code.  In addition to allowing for any character within the quotes,
these strings can span multiple lines.

The version has been bumped to 1.1.1 in __version__.py.

Support for getting the excerpt is also supported in the "getPost" option as
well as when converting a header object into a string.