Fixed Header __str__ Method
Actually, there was nothing wrong with the method per say.  The problem was that
we modified the header string when processing the header categories in
fileprocessor.py.  That processing removed information that was extraneous for
WP blog software, but useful for blogtool.  Now a separate category list is
generated  and returned for use with posting.

Also, now raise FileProcessorErrors where before we just quit.