Added Ability to Supply Custom More Text
By default, for posts that have part of their content hidden, WP displays a MORE
link at the bottom of the post.  This text can be customized by modifying the
HTML comment that generates the MORE link.  Simply add the custom text right
afte the 'more' in the HTML comment like so:  `<--more Custom More Text-->`.

This is now supported by detecting a single line in the post file that starts
with at least 3 '+' characters, which can be space separated.  If no text is
supplied afterwards, then the standard 'MORE' link is generated.  Any text that
does follow the '+' characters is then used as the text in the 'MORE' link.  Any
trailing '+' characters are ignored.

When getting the post from the blog, the `wp_more_text` field is now checked and
an appropriate line is generated for the post file using the new '+' character
notation.

The old `### MORE ###` is still supported for historical reasons.