Use Wordpress XMLRPC Methods for Posting
The `publishPost` and `editPost` methods in wp_proxy.py now use the native
Wordpress XMLRPC methods for posting.  To facilitate this, the new post data
class now builds a `wpStruct` in addition to the `metaweblogStruct`.

The old `buildPost` function is now gone, while the old `utils._convertTime`
function is now an internal method in the post class.  It is used to convert a
time string assignment to the `dategmt` attribute to an appropriate XMLRPC
DateTime object.